feat: July 23-24 closure wave -- live PQ handshake, transfer reliability, transactional updater, browser identity
Uncommitted canonical-tree work (115k insertions, 1245 files) now durably
recorded. Gate before commit: 8959 passed / 0 failed fast suite, ruff clean,
mypy clean (218 files), secrets sweep clean.

Security
- channel: v3 PQ hybrid daemon handshake (X25519 + ML-KEM-768) with signed
  canonical suite offer, full transcript binding, mutual key confirmation;
  native PQ capability advertised only after in-process encap/decap self-test;
  legacy channels rejected by default (explicit downgrade policy required)
- authenticated Double Ratchet cutover: CAPS frame is the exact final legacy
  boundary -- removes the old try-DR-then-legacy fallback (a downgrade surface)
- direction-separated native-transfer secrets kill AEAD key/nonce reuse at
  chunk index 0; facade refuses recreation under the same channel roots
- v1-HELLO downgrade telemetry bounded (pinned-only per-peer slots + saturating
  aggregate) -- closes a memory-DoS primitive; replay cache now lock-guarded
- key_material.py: fail-closed key authority persistence (TOCTOU-aware inode
  pinning, hard-link atomic first publication)

Reliability
- idempotent uploads: stable idempotency keys, durable staging before 202,
  exact replay after response loss; re-hash before phone retry replay
- authenticated durable FILE_COMMIT receipts prove receiver completion
- relay: exact 512 MiB payload budget + 4 MiB control reserve, drained shutdown
- chunk receipt replay, restart-safe resume ownership, bounded adaptive
  high-RTT windows (fixes 250 KiB/16-chunk BDP collapse on 596 ms routes)
- storage lifecycle: content-verified graph audit, recoverable quarantine,
  rollback, 30-day-grace purge

Platform
- transactional updater (update_transaction, standalone_updater, sigstore
  verification, external update helper)
- browser identity: self-contained argon2id + ed25519 WASM (CSP-safe),
  possession-based authority lifecycle, service-worker outbox durability
- inline PEP 561 stubs at native/one_link_native/ replace stubs/ tree
- ~100 new test files: adversarial lost-ACK, duplicate-init, concurrent
  finalization, crash recovery, relay overload, formal-model gates,
  product-copy truth gates

Docs de-hyped: alpha claims scoped to shipped reality; LT fountain codes
correctly attributed (RaptorQ deferred pending IPR review).

+++ native/ol_onion/tests/property_obfs.rs
@@ -1,4 +1,4 @@
-//! Property tests for the Row 7 transport_obfs primitive + handshake.
+//! Property tests for the Row 7 `transport_obfs` primitive + handshake.
 //!
 //! Two surfaces, two gate tiers:
 //!   - `primitive` (cheap, pure XOR): 1M iters CI default.
@@ -200,7 +200,7 @@ proptest! {
         client_seed in any::<[u8; 32]>(),
         server_seed in any::<[u8; 32]>(),
         now in any::<u64>(),
-        flip_idx in 0u8..HANDSHAKE_LEN as u8,
+        flip_idx in 0usize..HANDSHAKE_LEN,
     ) {
         let bridge = BridgeKeypair::from_parts(bridge_seed, bridge_id);
         let bridge_pk = *bridge.public.as_bytes();
@@ -209,7 +209,7 @@ proptest! {
 
         let client = ClientHandshake::start(&mut client_rng, &bridge_pk, &bridge_id, now);
         let mut tampered = *client.first_message();
-        tampered[flip_idx as usize] ^= 0x01;
+        tampered[flip_idx] ^= 0x01;
         let r = ServerHandshake::accept(&mut server_rng, &bridge, &tampered, now);
         // Either BadMac (most likely) or SmallOrderPubkey (vanishingly
         // rare; flipping byte 0 may yield a low-order point).
